requirements_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements_dev.txt requirements_dev.in
#
attrs==22.2.0
    # via pytest
autoflake==2.0.2
    # via -r requirements_dev.in
bakery==1.0.0
    # via -r requirements_dev.in
black==23.1.0
    # via -r requirements_dev.in
click==8.1.3
    # via black
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements_dev.in
mypy-extensions==1.0.0
    # via black
packaging==23.0
    # via
    #   black
    #   pytest
pathspec==0.11.1
    # via black
platformdirs==3.1.1
    # via black
pluggy==1.0.0
    # via pytest
pyflakes==3.0.1
    # via autoflake
pytest==7.2.2
    # via
    #   -r requirements_dev.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements_dev.in
